% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTS.R
\name{assign_per_bin}
\alias{assign_per_bin}
\title{for switchers in each bin, assign those switching time to the right number of continuers (among all continuers), then return the assigned continuers, the unassigned continuers for the next round of assign. Early bins is more likely to be assigned than later bins}
\usage{
assign_per_bin(counts, bin, swi_time = swi_time, cens_time = cens_time)
}
\description{
for switchers in each bin, assign those switching time to the right number of continuers (among all continuers), then return the assigned continuers, the unassigned continuers for the next round of assign. Early bins is more likely to be assigned than later bins
}
