% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTS.R
\name{sim_cov}
\alias{sim_cov}
\title{simulate covariates, including ids and switching variable}
\usage{
sim_cov(n = n, prob = 0.3)
}
\arguments{
\item{n}{Total sample size for both continuers and switchers combined}

\item{prob}{probability of switchers}
}
\value{
a list of data frames, respectively for covariate for continuers, switchers, and continuers and switchers combined
}
\description{
simulate covariates, including ids and switching variable
}
