% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTS.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data}
\usage{
sim_data(
  n = 5000,
  seed = 123,
  swi_min = 1,
  swi_max = 1.5,
  wshape = 1.5,
  wscale = 0.05,
  hr_bf = 2,
  hr_af = 0.5,
  log_hr_confound = 1,
  t_fow = 6,
  swi_prob = 0.3
)
}
\arguments{
\item{n}{Total sample size for both continuers and switchers}

\item{seed}{random seed}

\item{swi_min}{on a scale of 0 which is the treatment initiation time, to follow up end, when does the treatment switching starts to occur}

\item{swi_max}{on a scale of 0 which is the treatment initiation time, to follow up end, when does the treatment switching stops to occur}

\item{wshape}{the shape parameter for weibull distribution that is used to simulate the baseline hazard}

\item{wscale}{the scale parameter for weibull distribution that is used to simulate the baseline hazard}

\item{hr_bf}{hazard ratio before switching/pseudo-switching}

\item{hr_af}{hazard ratio after switching/pseudo-switching}

\item{log_hr_confound}{hazard ratio due to the events before switching/pseudo-switching as confounding factor}

\item{t_fow}{the maximum follow up end time}

\item{swi_prob}{probability for the switching variable to take value of 1}
}
\value{
a data frame contining the continuers and switchers, where the variable swi = 0 indicate continuers, swi = 1 indicate switchers
}
\description{
Simulate data
}
