% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTS.R
\name{get_count_per_bin}
\alias{get_count_per_bin}
\title{cut switching time into bins, and record the numbers
return a list containing the switchers, continuers, the cut of the bins, number of switchers and continuers in each bin}
\usage{
get_count_per_bin(swi, cont, breaks, swi_time = swi_time)
}
\description{
cut switching time into bins, and record the numbers
return a list containing the switchers, continuers, the cut of the bins, number of switchers and continuers in each bin
}
